proto: move serial_message from xfer to op
