+++ crates/client/src/client/dnssec_client.rs
@@ -15,11 +15,9 @@ use crate::client::Client;
 use crate::proto::ProtoError;
 use crate::proto::dnssec::DnssecDnsHandle;
 use crate::proto::dnssec::TrustAnchors;
-use crate::proto::runtime::{TokioRuntimeProvider, TokioTime};
 use crate::proto::op::DnsRequest;
-use crate::proto::xfer::{
-    DnsExchangeBackground, DnsHandle, DnsRequestSender, DnsResponse,
-};
+use crate::proto::runtime::{TokioRuntimeProvider, TokioTime};
+use crate::proto::xfer::{DnsExchangeBackground, DnsHandle, DnsRequestSender, DnsResponse};
 
 /// A DNSSEC Client implemented over futures-rs.
 ///
