+++ crates/proto/src/multicast/mdns_client_stream.rs
@@ -21,8 +21,9 @@ use crate::BufDnsStreamHandle;
 use crate::error::ProtoError;
 use crate::multicast::mdns_stream::{MDNS_IPV4, MDNS_IPV6};
 use crate::multicast::{MdnsQueryType, MdnsStream};
+use crate::op::SerialMessage;
 use crate::runtime::TokioTime;
-use crate::xfer::{DnsClientStream, SerialMessage};
+use crate::xfer::DnsClientStream;
 
 /// A UDP client stream of DNS binary packets
 #[must_use = "futures do nothing unless polled"]
