+++ crates/proto/src/op/serial_message.rs
