+++ crates/proto/src/tests/udp.rs
@@ -9,13 +9,13 @@ use std::{println, process, thread};
 use futures_util::stream::StreamExt;
 use tracing::debug;
 
-use crate::op::{DnsRequest, DnsRequestOptions, Message, Query};
+use crate::op::{DnsRequest, DnsRequestOptions, Message, Query, SerialMessage};
 use crate::rr::rdata::NULL;
 use crate::rr::{Name, RData, Record, RecordType};
 use crate::runtime::RuntimeProvider;
 use crate::udp::{UdpClientStream, UdpStream};
 use crate::xfer::dns_handle::DnsStreamHandle;
-use crate::xfer::{DnsRequestSender, DnsResponse, FirstAnswer, SerialMessage};
+use crate::xfer::{DnsRequestSender, DnsResponse, FirstAnswer};
 use crate::{ProtoError, ProtoErrorKind};
 
 /// Test next random udpsocket.
